% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alert.R
\name{get_alert}
\alias{get_alert}
\title{Allow users to get alerts of bus.}
\usage{
get_alert(bus)
}
\arguments{
\item{bus}{Bus number.}
}
\value{
Dataframe with bus number and its corresponding alert information.
}
\description{
This function allows users to get alerts information based on the specified bus number.
}
\examples{
get_alert(bus = 8)
}
