% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_direction.R
\name{get_direction}
\alias{get_direction}
\title{Allow users to get direction based on bus number.}
\usage{
get_direction(bus, key = Sys.getenv("BUS_CLIENT_ID"))
}
\arguments{
\item{bus}{Bus number.}

\item{key}{Key for accessing the API.}
}
\value{
Dataframe with bus number and its corresponding direction.
}
\description{
This function allows users to look up the direction of a specified bus number.
}
\examples{
get_direction(bus = 156)
}
