% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_route.R
\name{get_route}
\alias{get_route}
\title{Allow users to get route names.}
\usage{
get_route(bus = NULL, key = Sys.getenv("BUS_CLIENT_KEY"))
}
\arguments{
\item{bus}{Bus number.}

\item{key}{Key for accessing the API.}
}
\value{
Dataframe with bus number and its corresponding route name.
}
\description{
This function allows users to get route names based on the specified bus number.
If bus number is not specified, it will provide the all bus number with its corresponding route names.
}
\examples{
get_route()
get_route(bus = 5)
}
