% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupstop.R
\name{lookupstop}
\alias{lookupstop}
\title{Allow users lookup the information of bus stop.}
\usage{
lookupstop(bus, stopname = NULL, dir = NULL,
  key = Sys.getenv("BUS_CLIENT_ID"))
}
\arguments{
\item{bus}{Bus number.}

\item{stopname}{Name of bus stop.It is optional and case insensitive.}

\item{dir}{Direction of routes.It is optional and case insensitive.}

\item{key}{Key for accessing the API.}
}
\value{
Dataframe with potential matching stop with bus number, stop ID, direction, longitude and latidude columns.
}
\description{
This function allows to lookup the information of stops,
including stop ID, direction, longitude and latitude based on
the arguments of bus number, stop name and stop direction.
}
\examples{
lookupstop(bus = 1, stopname = "Michigan")
lookupstop(bus = 1, dir = "Southbound")
lookupstop(bus = 1, stopname = "indiana", dir = "northbound")
}
